pyproject.toml
[project]
name = "procedural-knots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "glfw>=2.10.0",
    "imagecodecs>=2025.11.11",
    "ipdb>=0.13.13",
    "ipython-genutils>=0.2.0",
    "ipyvolume>=0.6.3",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.2",
    "napari>=0.6.6",
    "noise>=1.2.2",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "openmesh==1.2.1",
    "perlin-numpy",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "pyopengl>=3.1.10",
    "pyrr>=0.10.3",
    "pyvista>=0.46.4",
    "scipy>=1.16.3",
    "tifffile>=2025.10.16",
    "vedo>=2025.5.4",
]

[tool.uv.sources]
perlin-numpy = { git = "https://github.com/pvigier/perlin-numpy" }
